Center window at sample 0 when performing FFT, not on GPU
We want constructive interference at the right of the GPU plot (0 radians),
so we need to rotate the FFT by [k] -> (-1)^k.
This was previously done by having the GPU rotate the plot by 180 degrees.
However, phase-derivative mode does not need rotation on the GPU.
So it makes sense to rotate non-phase-derivative mode on the CPU.