Compute phase derivative between non-overlapping windows
Phase used to be subtracted between the two most recent spectra.
If redraw interval was shorter than FFT length,
the most recent spectra would overlap in time and be correlated
regardless of the harmonic content of the input,
resulting in phase artifacts.

Now we keep a history of spectra such that the newest and oldest
do not overlap in the time domain.
So it is now safe to use a redraw interval shorter than FFT length.